+++ js/constants.js
@@ -1,0 +1,64 @@
+const TITLE_LIST = [
+  'Уютная квартира',
+  'Раскошный коттедж',
+  'Элитная квартира',
+  'Светлая комната',
+  'Спальное место',
+  'Квартира с красивым видом',
+  'Место где можно приткнуться'
+];
+const TYPE_LIST = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel'
+];
+const CHECKIN_LIST = [
+  '12:00',
+  '13:00',
+  '14:00'
+];
+const CHECKOUT_LIST = [
+  '12:00',
+  '13:00',
+  '14:00'
+];
+const FEATURES_LIST = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner'
+];
+const DESCRIPTION_LIST = [
+  'с видом на парк',
+  'прекрасное место для романтических свиданий',
+  'тихое и уютное место',
+  'на оживленной улице',
+  'с видом на достопримечательности',
+  'уютный внутренний двор',
+  'с выделенным местом для гриля'
+];
+const PHOTOS_LIST = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+];
+const AVATAR_ADRESS = 'img/avatars/user/';
+const AVATAR_PHOTO_FORMAT = '.png';
+const ANNONCEMENT_COUNT = 10;
+
+export {
+  TITLE_LIST,
+  TYPE_LIST,
+  CHECKIN_LIST,
+  CHECKOUT_LIST,
+  FEATURES_LIST,
+  DESCRIPTION_LIST,
+  PHOTOS_LIST,
+  AVATAR_ADRESS,
+  AVATAR_PHOTO_FORMAT,
+  ANNONCEMENT_COUNT
+};
